sumup/api/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
db-dtypes
scikit-learn
